This commit fixes loadufs command line to take parameters: bin/tachyon loadufs <TachyonPath> <UnderfsPath> [<Optional Excluding Prefixes;>]
(1) It adds a new method UnderfsUtil.getInfo/5 for implementing the new loadufs command line while keeping the old getInfo/4 for compatibility.
(2) It adds toString methods in PrefixList and TachyonFS for printing meaningful debug messages.
(3) It fixes list() in UnderFileSystemSingleLocal to return absolute paths, to be consistent with UnderFileSystemHdfs.
(4) It smooths out the uses of list() in other modules.
(5) It fixes docs/Syncing-the-Underlying-Filesystem.md accordingly.