Remove the dependency on ant.
The only thing Tachyon uses from ant is LazyFileOutputStream.  Since
this is under the Apache license, it's easier to just include that one
file than to depend on ant.  Depending on ant means that (for example)
Spark jobs get ant jars in their CLASSPATH when using Tachyon, which
they may not want.