Additional system specific settings for building on Mac
Since Tachyon writes a lot of short lived files to the local file system
during the unit tests if the Spotlight indexing processes are not set to
ignore the Tachyon working copy trying to run the tests can hang
perpetually as Spotlight tries to re-index the working copy.  Adding the
Tachyon directory to the Spotlight exclusions list resolves this
problem.