Fix broken shade plugin execution
Previous commits broke shade plugin execution because they suppressed
generation of the empty JAR causing the shade plugin to complain.
Changing the packaging of the client module to pom resolves this and is
actually more appropriate packaging than JAR anyway.  This allows the
shade plugin to generate a valid uber-jar

Also applies similar changes to the assembly module which is used only
to build the distribution packages.  Again we change the packaging to
POM which removes the spurious warning about an empty JAR.

By changing the packaging of these modules to POM the jar plugin
override is no longer necessary and is removed