Propogate debug mode through sub-shells
The Tachyon scripts launch various sub-shells both locally and remotely.
If you are trying to debug why Tachyon is not launching properly the
usual thing to do is to launch with `bash -x` to get debugging output,
however this does not get propagated to sub-shells so sometimes if the
error is in another script launching with debugging won't tell you
anything useful.

This commit changes all the scripts to check whether debugging is
enabled and if so will propagate that through to the sub-shells by
launching them with `bash -x`