Note which user we're using to over SSH
Improves the echo statement used in tachyon-workers.sh to note which
user we are SSH'ing as which can be useful if you are experiencing
authentication failures because it can be unclear which password you are
expected to supply.  This is especially true if you have launched the
scripts using `sudo` as the password SSH expects you to provide is
likely different from the password you provided to sudo initially.