Improve how old RAM FS is unmounted on OS X
On OS X a RAM FS is used and each time Tachyon is started the old RAM FS
(if any) is unmounted and then a brand new one is formatted and mounted.
However because of the way in which the RAM FS is unmounted we end up
endlessly incrementing the device ID so after you've run Tachyon a bunch
of times you end up with a device ID like `/dev/disk22` and a huge list
of unused disks if you run `diskutil list`

This commit changes how the old RAM FS is unmounted to instead determine
the actual device ID behind the RAM FS volume and if present detach it
rather than unmounting it.  This causes OS X to unmount and throw away
the disk so it doesn't clutter the list of devices and we simply reuse
the device ID when we create the new disk rather than endlessly
incrementing it.