Move HDFS specific tests to underfs-hdfs module (TACHYON-221)
This commit moves the optional profile which runs the unit tests against
the HDFS under file system implementation into the underfs-hdfs module.

Doing this requires some Maven trickery to work correctly, we need to
depend on the tests jar of the core module as a test scope dependency,
pull in libthrift as a test scoped dependency and then use the surefire
plugins <dependenciesToScan> functionality to discover the existing
tests.

With this done running the -PhdfsTest profile will run the HDFS tests in
addition to the core unit tests since these tests are now specific to
the module.  This has the additional advantage that we can run tests over
multiple under file system implementations within the same build.