Move HDFS specific tests to underfs-hdfs module (TACHYON-221)
This commit moves the optional profile which runs the unit tests against
the HDFS under file system implementation into the underfs-hdfs module.

Doing this requires some Maven trickery to work correctly, we need to
depend on the tests jar of the core module as a test scope dependency,
pull in libthrift as a test scoped dependency and then use the surefire
plugins <dependenciesToScan> functionality to discover the existing
tests.

With the tests for HDFS moved to running from the underdfs-hdfs module
the initial runs showed that not everything was completely set up for
this test set up to run successfully, this fixes some issues with this:

- Adds missing necessary test scope dependencies
- Improves the factory tests so that the test that assumes it runs only
  in the core module explicitly assumes that