Enable HDFS tests by default (TACHYON-221)
This commit removes the specific test profile for the underfs-hdfs
module and instead simply sets up the POM such that the unit tests with
HDFS as an under file system are run by default.

As part of this commit the test cluster code for HDFS moves to the
underfs-hdfs module.  Also a remaining faulty test assumption in the
factory tests was addressed.