Continued test profile refactoring (TACHYON-221)
The localTest profile in the core module is no longer directly relevant
since it will always be used regardless as the core module now only
provides the local under file system implementation.

The glusterfsTest profile is moved to the underfs-glusterfs module.
Unlike the HDFS tests this remains a profile since Gluster FS requires a
proper cluster setup and can't be simulated.