Make Gluster FS profile runnable
This commit ports roughly the same POM changes as were introduced to the
underfs-hdfs module into the underfs-glusterfs module.  As before these
changes allow this module to run the unit test suite using a different
under file system - in this case Gluster FS.  Whether these changes are
sufficient for the tests to complete successfully has yet to be
verified.