Make underfs modules actually test the underfs (TACHYON-221)
In debugging the new Maven configurations that are supposed to run the
unit test suites using alternative under file system implementations I
discovered that the profiles as specified did not actually test the
under file systems.  The test code uses a JVM system property to pass in
the desired UFS implementation to use but the Maven profiles was setting
only a Maven property NOT a JVM property so the tests would look like
they were running and passing but weren't actually testing the right
file system.  This worked under the old code layout because that
explicitly passes the Maven property through to the forked JVM using the
Surefire configuration.

Therefore this commit adjusts the surefire configuration for
these modules appropriately such that they really do test the desired UFS
implementation.  The code that initializes the UFS to use now also
clearly prints which implementation it has selected which is useful for
sanity checking the configuration

As a result of this change I also discovered that many of the tests
assumed that they would always successfully get a cluster instance and
as such the @After methods of many test classes would throw spurious
NPEs when trying to run the tests with an improperly configured UFS
implementation.  As such many test classes are touched by the commit to
make them check that the cluster is non-null plus any other things they
are going to try and clear up that may be null if the cluster is
null