Upgrade JUnit and PowerMock to the latest versions
Currently JUnit 4.10 is used which is from Sept 2011 whereas the latest
version is 4.12 from Dec 2014.  The old version is licensed under the
Common Public License 1.0 which is a deprecated license that was
superseded by the Eclipse Public License which makes it problematic from
a licensing perspective.  Newer versions of JUnit are licensed under the
Eclipse Public License.

This commit upgrades JUnit to 4.12 and changes the hard coding of the
JUnit version in the dependency declaration to use a Maven property
to make this easy to change in the future.  Powermock was also upgraded
to the latest version in order to align with the JUnit upgrade