Add bundled code to LICENSE file
Tachyon bundles several third party libraries as-is in its
web-application.  These include all/part of the following:

- Bootstrap 2 (Apache License)
- jQuery (MIT License)
- Cookies.js (Public domain)
- Modernizr (MIT License)
- Pygments (BSD 2-Clause License)

Entries for each of these are added to the LICENSE file along with the
relevant licenses

The NOTICE is also updated to update the copyright year to 2015 and to
inform users that some portions of Tachyon are under alternative
licenses and they should consult LICENSE to see the details of this.