Fix TachyonURI authority parsing (TACHYON-289)
TachyonURI parses URIs from strings incorrectly in cases where the
authority does not contain a port.  Since the port is an optional
component of the authority (see RFC 3986 Section 3.2) this is incorrect.
This wrong assumption means that Tachyon can treat some URIs as not
having authorities causing them to get mangled and produce odd error
messages as reported in TACHYON-289

This commit corrects the parsing of the authority component and updates
the corresponding tests to cover authorities without ports and correct
existing tests that were using this wrong assumption.

NetworkUtils.replaceHostName() is also appropriately updated to not
include a port if there wasn't one to start with