Tachyon command should abort tests early if they receive a signal
Currently if you use `bin/tachyon runTests` and for whatever reason it
starts failing e.g. misconfiguration using Ctrl+C or kill to terminate
it only terminates the current test but still proceeds to run the
subsequent tests.

This commit changes the logic of both runTests and runTest so that if
the JVM receives a signal (whether via Ctrl+C, kill or some other means)
that causes further testing to be aborted.