Revamped storageId handling on master and client
This commit adds storage id information to the thrift metadata, so that
clients can use the information to query workers in the best order. It
also lets the master handle having multiple storage directories per
worker. While currently we don't store a block on multiple storage
directories on a worker, we anticipate this to happen in the future, and
right now it uses minimal overhead.

The main changes are to the thrift metadata, where we couple each worker
address with a list of storage ids. In turn, we modified the BlockInfo
class to handle this new metadata representation, and we also modified
the worker storage code to explicitly report removing blocks from a
certain storageId.

We also modified the RemoteBlockInStream class to take advantage of this
new metadata. Now it will sort the workers by their storage id tier, so
when we need to read part of the block remotely, we can query workers in
order of where they have stored the block.

Conflicts:
	core/src/main/java/tachyon/master/BlockInfo.java
	core/src/test/java/tachyon/master/BlockInfoTest.java