Parameterize Perf slave log file names
When running Tachyon perf on a shared file system all the slaves log to a single
log file slaves.log which means their logging is interleaved and hard to read.

This commit parameterizes the name of the slaves log file to add the slave ID to
the log file name e.g. slave-1.log