Initial support for reserving multiple blocks during writing
When the block size is smaller than the file size the write performance
can be significantly reduced because the client keeps having to go back
to the master to obtain new block IDs.  This commit adds a framework for
allowing users to configure reserving multiple blocks at a time.  When
configured the client will reserve N blocks in a single request to the
master and won't need to go back to the master until those N blocks are
used up.

If the client ends up needing less blocks than were allocated the extra
blocks are cancelled when the write is completed.