Add a script that helps with scripting configuration of test cases
Test cases as they stand are configured via XML files which while nice
for humans to edit are something of a pain to script against.  This
commit adds a script that levarages the `xsltproc` command line tool to
automatically transform a configuration file.

The intent of the script is to make it easy for end users to script up
their benchmarks so that they can script a set of runs with different
configuration parameters without having to manually edit the XML files
between each run.