Add auto-refresh functionality to Web UI
This commits adds auto-refresh functionality to the Web UI which is
useful if you are keeping an eye on Tachyon's state and don't want to
have to hit Refresh manually.

Auto-Refresh is off by default and is toggled on via a link in the top
navigation.  Once enabled the page will refresh after 15 seconds and
auto-refresh remains enabled on page reload.

Refresh uses `window.location.replace()` so it does not fill the users
history and the back button will work normally

As part of this commit the JSPs were refactored slightly to pull the
script definitions out into includeable JSPs.  One for scripts that need
to appear at the start of a page and one for scripts that must appear at
the end of the page.