Make local integration tests a profile (TACHYON-49, TACHYON-221)
Per PR comments it can also be useful to run parallel builds each
testing a different Under FS implementation.  To this end the local
integration tests are placed into a profile which is active by default.
This means that normal builds will always include this but in
environments where you want to run only a specific integration tests
they can optionally be disabled via the the maven !profile syntax