Made TachyonFileSystem methods throw better exceptions
We were turning everything into an IOException, which can make it hard
for callers to determine what went wrong in an operation. So now the
TachyonFileSystem and related classes don't wrap exceptions in an
IOException.

Also the FileSystemMasterClient uses an exponential backoff retry,
rather than looping forever, before throwing an exception and exiting.

There are a lot of miscellaneous refactors for deprecated code and tests
to get this to compile.