[TACHYON-993] Implement a better exists method
Currently, to check if a file exists, users have to call getInfo on the
file, and catch if it throws an exception. Instead, we changed
TachyonFileSystem.open, so that when a user opens a file, we return null
if the file was not found. Thus users can check if a file exists by
trying to open it.