[TACHYON-993] Added openIfExists method
Most people expect open to throw an exception if the path does not
exist, but currently it would returns null, so that people could use it
as an exists method.

Rather than have an exists method, which introduces atomicity
problems (the file could be deleted in between checking for existence
and opening), we added an openIfExists method, which has the same
behavior as the old open method. The new open method will throw an
exception if the path does not exist.