[TACHYON-724] Allow for Configurable Storage Levels
We implement configurable storage levels, as described in the JIRA.

Most of the changes involve modifying the codebase to identify a storage
level by its alias name, as opposed to its enum value (which we don't
have anymore), or its level in the hierarchy (which can change across
workers).

We get rid of the StorageLevelAlias class, and replace it with
StorageTierAssoc, which maps between aliases and ordinal tier values, on
a specific worker, or on the master (for the global ordering of storage
tiers).

Also, we get rid of the confusion between the storage tier enum value
and storage level, and only use the ordinal number to sort storage
tiers. This simplifies classes like StorageDirId.