[TACHYON-1167] Significantly speed up integration tests
The default Thrift TThreadPoolServer timeout is 60 seconds. This was causing our
integration tests to wait until all outstanding RPCs were resolved before
completing. As a result, we were spending half of our integration test time
waiting for the Thrift server to stop. This commit fixes this by making the
timeout configurable via TachyonConf, and setting it to 0 for tests.