[TACHYON-988] Added LocalTachyonClusterResource
This is a Junit rule, which serves as sort-of a mixin for integration
tests. The goal is that instead of creating and managing your own
LocalTachyonCluster for every integration test, you use a
LocalTachyonClusterResource, with some basic defaults, and can employ
per-class or per-method configuration on it.

The rule can also be created as a static class, which means its only
initialized once per suite, rather than once per method.