[TACHYON-1333] Fixed UFS mkdir persistence
One case where loadMetadata would fail is when the user calls "ls" on a
directory in the UFS that hasn't been loaded yet, which creates the
inode in the tree, but doesn't mark it as persisted, because the UFS
failed to create the directory (since it already existed). Then when
they call loadMetadata, creating the directory would again try to mark
it as persisted and fail for the same reason, and loadMetadata would
crash because it expected that either a directory was created, or an
exception is thrown.

This modifies InodeTree.createPath to mark a directory as persisted
regardless of whether the ufs directory was actually created or
not (because if it was not, that means it already existed). Now
loadMetadata will throw the proper error message when the directory has
already been loaded into Tachyon.

Added a test in TfsShellTest that reproduces the error.