Made loadMetadata more robust
It's still possible to construct scenarios where the directory has an
entry in InodeTree, and exists in the UFS, but is not marked as
persisted. So to be completely robust, we check all the results of the
createPath for the created inode id, and throw
FileAlreadyExistsException if the traversal did nothing.