Shell code update
I introduced a new checking of the input parameters, so if program execute without parameters or having three or more input parameters, it will execute printUsage function and exit with error code 1. Also, I changed the IF-ELSIF-ELSE statements for a CASE which turns to be more clean and have a better readable.