[SMALLFIX]Convert semaphore to unfair
If a thread read, then it hold the read lock and don't unlock, then another thread want to write lock, as we all know, this thread will never get the lock, but a third thread want to get the read lock, it also never get the lock if we set the semaphore to `fair`. So i set it to `unfair` to let other read thread get read lock.