Modify the logic in the assignment of ALLUXIO_LOGS_DIR.
The ALLUXIO_LOGS_DIR variable is used by Alluxio to determine the
location of log files. The user can use environment variable and
alluxio-site.properties file to configure this. Here is how the location
is determined.
1. If an environment variable named ALLUXIO_LOGS_DIR is set (preferrably
in alluxio-env.sh), then this locatio is used.
2. If no such environment variable is set, then the property called
alluxio.logs.dir in alluxio-site.properties is used to determine the
location.
3. Finally, if neither $ALLUXIO_LOGS_DIR environment nor
alluxio.logs.dir is set, then the default location $ALLUXIO_HOME/logs is
used to store the logs.