Ensure order of audit log entries.
Accesses to the same metadata are preserved with inode lock.
To do so, as soon as we enter the critical section protected by inode
lock, append the audit entry to the queue and mark it as uncommitted,
i.e. not ready to be written to audit log.
In the critical section, after obtaining the information, e.g. file
owner, group and permission, set the fields of the rpc context and mark
it as committed, indicating that this entry can be written to audit log.