Modify the exception handling during thread wait.
When a thread is waiting for available space from the queue
to insert audit log messages, it can get interrupted by another thread.
This is captured by the InterruptedException. When this happens, it
usually means there is something wrong with the system and there is no
point to continue waiting. Therefore, we clear the interrupted flag and
return.

Make member variables of MasterAuditContext private.