Handles recursive file and directory operations.
AuditContext should not cause FileSystemMaster to throw
FileNotExistException. Instead, handle the recursive and non-recursive
operations separately.

For non-recursive operations, AuditContext should set its srcInode to be
the parent of the target inode. As said, AuditContext should not cause
FileNotExistException. Therefore, even if the parent of target inode
does not exist, it should not throw exception. Consequently, previous
getParentInodeDirectory does not meet the requirement. That's why we add
peekParentInodeDirectory.

For recursive operations, AuditContext should not cause
FileNotExistException. If the parent of target inode does not exist,
then AuditContext should set srcInode to the last existing inode
(whether file or directory) in the LockedInodePath.