Remove retry logic.
If the thread pool throws a RejectedExecutionException, we throw a
RuntimeException to indicate that an increase in thread pool size is
necessary.

If the thread pool is interrupted while waiting for its worker threads
to terminate, we also throw a RuntimException because something wrong
has happened somewhere else that causes the main log server thread to
terminate.