Refactor AlluxioRemoteLogFilter.
Rename the option from AppenderName to ProcessType. It seems more
intuitive for a process to pass to a remote log server its own process
type than appender name. Furthermore, if we stick to AppenderName, it
will likely lead people to think that the value of mAppenderName has
to match the name of the appender in log4j.properties. It is not the
case. Here the information we attache to the MDC can be anything, as
long as the log server and client have an agreement on how to process
LoggingEvents with different such information.