[ALLUXIO-3120] Fix issue where secondary deletes completed log  (#6930)
* Fix issue where secondary deletes completed log

Before it was possible for a shutting-down secondary
to try completing its current log after the primary
has already completed the current log. In this case,
the secondary would delete the completed log before
realizing that the current log no longer exists (the
primary has already renamed it). With this PR, we
will never delete completed logs.

There is still a race where a secondary could
delete an empty current log that the new primary
has created. This will be addressed in a separate
PR.

* add javadoc