Update kubernetes/conf/alluxio.properties.template
In the default integration/kubernetes/conf/alluxio.properties.template the address for  ALLUXIO_WORKER_DATA_SERVER_DOMAIN_SOCKET_ADDRESS is not complete. This is properly explained in https://www.alluxio.org/docs/1.7/en/Running-Alluxio-On-Docker.html for enabling short circuit reads in alluxio worker containers using unix domain sockets. 
Just because of a missing complete path for unix domain socket the alluxio worker was not able to come up in kubernetes when short circuit read was enabled for alluxio worker.