overloading for LocalAlluxioMaster
Added a new create method which only takes a workDirectory argument. This version of the method is only used in the MultiMasterLocalAlluxioCluster and helps us avoid some Nullable annotations.

I've also annotated the ServerSocket fields for rpc and web as Nullable because when creating the master either could be null which will stop the master process from using the specified sockets for the master.