Add gRPC authentication timeout
Previously there was no timeout on the channel authentication. During testing if a channel got hung up during authentication then we would never return and tests would fail or time out. By adding a timeout we can force retries within a specified period of time.

I've added the property key for alluxio.master.grpc.channel.auth.timeout so that we can control this timeout value. For testing utilities I've set the value to 2sec. Default is at 30sec which is half of the 60sec channel shutdown.

If more modification of the channel shutdown timeout is desired we should address it in a spearate PR.