[SMALLFIX] - Master Process refactor - Use sockets to temporarily reserve ports for testing (#8311)
* refactor master processes use sockets to reserve ports

- Reduced some code duplication between the Alluxio Master and Job Master.
- Added ServerSocket arguments to "reserve" ports. This is useful for testing because it reduces the chance of race conditions where mulltiple processes attempt to use the same port.
- Added reserving sockets to integration tests which need both AlluxioMaster and JobMaster

* Add license header to AlluxioMasterProcess

* Use the correct javadoc link for MasterProcess

* Use a list of valid services instead of a mapping

Each service already defines the port mapped to the service, so we don't need to redefine the mapping. Simply calling service.getPortKey() returns the proper key.

Also made sure there is only one list of valid services so that we won't have to maintain two lists.

* Use immutable list of services and respect bind hosts

Use the right bind host for Master process sockets.

Also make findbugs happy by using an immutable list from guava instead of a normal list.

* Remove JobMasterProcess

JobMasterProcess is unnecessary and adds complexity where it is not needed. Simply letting AlluxioJobMasterProcess is enough.

* Fix rest API tests that use job master

* fix for journalFencingFailover test

* Master process clean up

- Renamed internal variable rpcBindSocket inside of setupBindSocket to bindSocket. Renamed because more than just rpc sockets are handled in the method.

- Removed exception handling in the getAddressXfromBindSocket. It makes error propagation simpler.

* Decrease timeouts for embedded journal integration tests

- Allows us to set the timeouts after initilization on the GrpcManagedChannelPool. Before the timeouts were only set from site properties or default values - this meant we couldn't modify them for testing. The embedded journal test would fail because we were waiting too long for channels to timeout.

The method is marked as VisibleForTesting because the only time we should need to change the channel pool is if we are testing.

I also added a timeout to the mAuthenticated.get line in the SaslStreamClientDriver so that we don't hang indefinitely waiting for authentication.

* Cleanup  master processes

- Changed a few property keys in JobMasterProcess which needed to be for Job master instead of master.
- Removed use of the deprecated Throwables.propagate (see https://github.com/google/guava/wiki/Why-we-deprecated-Throwables.propagate)

* overloading for LocalAlluxioMaster

Added a new create method which only takes a workDirectory argument. This version of the method is only used in the MultiMasterLocalAlluxioCluster and helps us avoid some Nullable annotations.

I've also annotated the ServerSocket fields for rpc and web as Nullable because when creating the master either could be null which will stop the master process from using the specified sockets for the master.

* MasterProcess javadoc updates

- Updated some javadoc descriptions for MasterProcess.
- Marked startServing as a protected instead of package-private

* Add gRPC authentication timeout

Previously there was no timeout on the channel authentication. During testing if a channel got hung up during authentication then we would never return and tests would fail or time out. By adding a timeout we can force retries within a specified period of time.

I've added the property key for alluxio.master.grpc.channel.auth.timeout so that we can control this timeout value. For testing utilities I've set the value to 2sec. Default is at 30sec which is half of the 60sec channel shutdown.

If more modification of the channel shutdown timeout is desired we should address it in a spearate PR.

* Checkstyle fix for grpc authentication

* Uncomment testConfigurationDefault for auth timeout

* Cleanup of MasterProcess and subclasses

MasterProcess.java
- Removed unused check for connectionTimeout in MasterProcess
- Updated a small bit of javadoc in MasterProcess
- Simply set configuration by passing the port instead of calling Integer.toString

AlluxioMasterProcess.java and AlluxioJobMasterProcess.java
- Move factory to bottom of class

LocalAlluxioCluster.java
- Add no-arg constructor