refactor FileSystem`get` to `create` and implement caching for `
Any method which desires a custom configuration for an Alluxio client should be calling `create` for FileSystem client instances. Otherwise if the configuration from system properties and site properties is desired, then the user can call `FileSystem.get` in order to get an instance of a client.

Calling `get` will either create a new client object and cache it or it will return a cached client object` This stops us from creating too many thread pools in the underlying FileSystemContext that the BaseFileSystem uses.

We use a key based upon the client's security subject and the authority (master connection details). This should suffice for most situations.