[SMALLFIX] Create clients less often in REST handlers (#8329)
* Create clients less often in REST handlers

The web server rest handler instances are created every time a request
is propagated. We shouldn't be creating new client objects within the
service handlers.

Instead we initialize a client in the servlet's `init` method which will
only be called once. Then subsequent calls to the REST service handler
wilk only utilize the single client.

* Update resource key names

It's good to be consistent.